+++ src/aks-preview/azext_aks_preview/_help.py
@@ -218,12 +218,19 @@ helps['aks create'] = """
               Using together with "azure" network plugin.
               Specify "azure" for Azure network policy manager and "calico" for calico network policy controller.
               Defaults to "" (network policy disabled).
+        - name: --network-dataplane
+          type: string
+          short-summary: The network dataplane to use.
+          long-summary: |
+              Network dataplane used in the Kubernetes cluster.
+              Specify "azure" to use the Azure dataplane (default) or "cilium" to enable Cilium dataplane.
         - name: --enable-cilium-dataplane
           type: bool
           short-summary: Use Cilium as the networking dataplane for the Kubernetes cluster.
           long-summary: |
               Used together with the "azure" network plugin.
               Requires either --pod-subnet-id or --network-plugin-mode=overlay.
+              This flag is deprecated in favor of --network-dataplane=cilium.
         - name: --no-ssh-key -x
           type: string
           short-summary: Do not use or create a local SSH key.